+++ src/@types/index.ts
@@ -1,3 +1,3 @@
-export * from "./volunteer-card-item";
+export * from "./volunteer-get-type";
 export * from "./ordering-type";
 export * from "./volunteer-post-type";
